moved scidb structures to a separate file, implemented properties selection
