+++ src/shim_client_structs.h
@@ -6,6 +6,8 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/lexical_cast.hpp>
 #include <boost/assign.hpp>
+
+
 using namespace boost::assign;
 
 namespace scidb4gdal {
@@ -15,15 +17,15 @@ namespace scidb4gdal {
   };
 
   enum Properties {
-    SRC_WIN
+    SRC_WIN, T_INDEX
   };
   
-  struct ImageProperties {
+  struct SelectProperties {
 	float src_coords [4]; //usually image coordinates in the order xmin,ymin,xsize,ysize
+	int temp_index;
 	
-	static ImageProperties *parsePropertiesString ( const string &propstr ) {
-	    std::map<string,Properties> propResolver = map_list_of ("src_win", SRC_WIN);
-	    ImageProperties *out = new ImageProperties();
+	static void parsePropertiesString (SelectProperties *properties,  const string &propstr ) {
+	    std::map<string,Properties> propResolver = map_list_of ("src_win", SRC_WIN)("t",T_INDEX);
             vector<string> parts;
             boost::split ( parts, propstr, boost::is_any_of ( ";," ) ); // Split at semicolon and comma for refering to a whole KVP
 	    //example for filename with properties variable    "src_win:0 0 50 50;..."
@@ -39,9 +41,13 @@ namespace scidb4gdal {
 			boost::split ( c, kv[1], boost::is_any_of ( " " ) ); 
 			
 			for (int i = 0; i < c.size(); i++) {
-			  out->src_coords[i] = boost::lexical_cast<float> ( c[i] );
+			  properties->src_coords[i] = boost::lexical_cast<float> ( c[i] );
 			}
 			
+			break;
+		      case T_INDEX:
+			  properties->temp_index = boost::lexical_cast<int>(kv[1]);
+			//TODO maybe we allow also selecting multiple slices (that will later be saved separately as individual files)
 			break;
 		      default:
 			continue;
@@ -49,11 +55,13 @@ namespace scidb4gdal {
 		  }
                 }
             }
-            
-            return out;
       }
     };
     
+    /**
+     * Structure to pass and store all the connection parameters defined in the connection string that
+     * was passed to a gdal function using the filename
+     */
     struct ConnectionPars {
 
         string arrayname;
@@ -62,7 +70,7 @@ namespace scidb4gdal {
         string user;
         string passwd;
         bool ssl;
-	ImageProperties *properties;
+	
 
         ConnectionPars() : arrayname ( "" ), host ( "https://localhost" ), port ( 8083 ), user ( "scidb" ), passwd ( "scidb" ) {}
 
@@ -72,7 +80,7 @@ namespace scidb4gdal {
             return s.str();
         }
 
-        static ConnectionPars *parseConnectionString ( const string &connstr ) {
+        static ConnectionPars *parseConnectionString ( const string &connstr, SelectProperties *properties ) {
 	    std::map<string,ConStringParameter> paramResolver = map_list_of ("host", HOST)("port", PORT) ("array", ARRAY) ("user",USER) ("password", PASSWORD);
             if ( connstr.substr ( 0, 6 ).compare ( "SCIDB:" ) != 0 ) {
                 Utils::error ( "This is not a scidb4gdal connection string" );
@@ -94,7 +102,7 @@ namespace scidb4gdal {
 	      int end = start + propToken.length();
 	      connectionString = astr.substr(0,start-1);
 	      propertiesString = astr.substr(end,astr.length()-1);
-	      out->properties = ImageProperties::parsePropertiesString(propertiesString);
+	      SelectProperties::parsePropertiesString(properties, propertiesString);
 	    } else {
 	      connectionString = astr;
 	    }
@@ -138,6 +146,19 @@ namespace scidb4gdal {
         }
 
 
+    };
+    
+    /**
+     * Helper structure filled while fetching scidb binary stream
+     */
+    struct SingleAttributeChunk {
+        char *memory;
+        size_t size;
+
+        /*
+            template <typename T> T get ( int64_t i ) {
+                return ( ( T * ) memory ) [i]; // No overflow checks!
+            }*/
     };
 }
 #endif
